pyproject.toml
[tool.poetry]
name = "dummy-model"
version = "0.1.0"
description = "Dummy Model"
authors = ["Rishabh Malviya <Rishabh.Malviya@docusign.com>"]
readme = "README.md"
packages = [{include = "dummy_model"}]

[tool.poetry.dependencies]
python = "3.10.6"
typer = "^0.7.0"
hydra-core = "^1.3.2"
pandas = "~2.1.4"
sagemaker = "2.191.0"
dictdiffer = "0.9.0"
ds-snowflake-utils = "1.2.0"
# optional/extra dependencies
pytest = { version = "^7.2.2", optional = true }
ipykernel = { version = "^6.23.2", optional = true }

[tool.poetry.extras]
dev = [
    "pytest",
    "ipykernel"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "docusign-internal-pypi"
url = "https://artifactory.docusignhq.com/artifactory/api/pypi/pypi-virtual/simple"
default = true
